Move files to match with new project folder structures.
