+++ apps/server/src/modules/idp-console/api/index.ts
@@ -1,1 +1,2 @@
+export * from './idp-sync-console';
 export * from './synchronization.uc';
