+++ apps/server/src/modules/idp-console/api/interface/system-type.enum.ts
