+++ apps/server/src/modules/idp-console/api/interface/users-sync-options.interface.ts
