+++ apps/server/src/modules/idp-console/api/loggable-exception/failed-update-lastsyncedat.loggable-exception.ts
