+++ apps/server/src/modules/idp-console/api/loggable-exception/synchronization-unknown-error.loggable-exception.ts
