+++ apps/server/src/modules/idp-console/testing/users-sync-options.builder.ts
@@ -1,5 +1,5 @@
 import { EntityId } from '@shared/domain/types';
-import { SystemType, UsersSyncOptions } from '../interface';
+import { SystemType, UsersSyncOptions } from '../api/interface';
 
 export class UsersSyncOptionsBuilder {
 	static build(systemType: SystemType, systemId: EntityId): UsersSyncOptions {
